Cleaned up text and made change in evolution-plot (Correct?)
I think the earlier plots are wrong, since they count the rows in the
data set and not the total RECOMPTE for each group. I left the other
plots in for now. Is this correct? The plots are not behaving exactly
how I would like, but did not figure out exactly why now.